requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-openai
langchain-community
faiss-cpu
pypdf
tenacity
pytest
httpx
tiktoken
streamlit
requests
python-dotenv